feature: add Auth redirect
