+++ package.json
@@ -7,6 +7,7 @@
     "fetch-jsonp": "^1.1.3",
     "firebase": "^5.5.5",
     "firebaseui": "^3.4.1",
+    "history": "^4.7.2",
     "lodash": "^4.17.11",
     "normalize.css": "^8.0.0",
     "react": "16.6.3",
